Break intersectionFunctions into separate files
